Is that professor rework i hear?

+++ src/IMU.h
@@ -2,9 +2,9 @@
 
 #include <Arduino_LSM9DS1.h>
 
-extern float cx, cy, cz;
-extern float crx, cry, crz;
+extern double cx, cy, cz;
+extern double crx, cry, crz;
 
 void initIMU(void);
 void calibrateIMU(void);
-void pollIMU(float& ax, float& ay, float& az, float& rx, float& ry, float& rz);
+void pollIMU(double& ax, double& ay, double& az, double& rx, double& ry, double& rz);
